Moved Jon's multinest test case into a separate test suite, with the test case disabled by default.
